whisper_requirements.txt
# Whisper Translation Service Requirements
# Core Whisper and audio processing
openai-whisper>=20231117
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.21.0
scipy>=1.9.0

# Translation services
googletrans==4.0.0rc1
gtts>=2.3.0

# Audio utilities (optional but recommended)
ffmpeg-python

# Additional utilities
requests>=2.28.0
